src: litra-toggle -> litra-oversight; fix if else
